[Game] Renamed _holdingContainer to HoldingContainer since it is now Public; [Game] Fixed setting the initial durability for a piece of equipment;
[Game] Fixed recognition/printing of items (SelectiveItems);
[Game] Fixed the operation of transfiguring things (ConvertItemLook);
[Game] Fixed repair of things;

[Game/Items] Clarifying values ​​for ItemTaskType;
[Game/Items] Changes to ItemUpdateRepair;
[Game/Items] Changes to ItemUpdateSecurity;
[Game/Items] Changes to EquipItem:
- added saving ImageItemTemplateId;

[Game/Quests] Changes to CompletedQuest:
- Id field in version 1.2 = UInt16, in version 5.0.7.0 = UInt32;

[Game/Skills] Added DynamicEffects cdass;
[Game/Skills/DynamicEffects] Added by BlessUthstin, SelectiveItems;
[Game/Skills/Effects/SpecialEffects] Fixed ItemConversion;
[Game/Skills/Effects/SpecialEffects] Fixed Skinize;

[Packets] Added CSBagHandleSelectiveItemsPacket with opcode 0x1A4;
[Packets] Fixed CSConvertItemLookPacket package;
[Packets] Added assumption about package naming to CSOffsets;

[Packets] Fixed SCCompletedQuestsPacket;
[Packets] Added SCGachaLootPackItemResultPacket package with opcode 0x100;
[Packets] Changes in the SCUnitState package, the flags field is also changed for Npc;

[SQL/Patches/Compact] Added patch file 2024-11-21-fix-skill_dynamic_effects.sql:
- In client version 5.0.7.0, TemplateId for items is encrypted in the `skill_dynamic_effects` table
into which the original item can be transformed.
- In version 4.0.3.6 this table is not yet encrypted, but it is slightly smaller than in version 5.0.7.0.

Attention! Be sure to apply patch 2024-11-21-fix-skill_dynamic_effects.sql to the compact.sqlite3 server database.

+++ AAEmu.Game/Models/Game/Char/CharacterQuests.cs
@@ -28,13 +28,13 @@ public class CharacterQuests
 
     private Character Owner { get; set; }
     public Dictionary<uint, Quest> ActiveQuests { get; }
-    private Dictionary<ushort, CompletedQuest> CompletedQuests { get; }
+    private Dictionary<uint, CompletedQuest> CompletedQuests { get; }
 
     public CharacterQuests(Character owner)
     {
         Owner = owner;
         ActiveQuests = new Dictionary<uint, Quest>();
-        CompletedQuests = new Dictionary<ushort, CompletedQuest>();
+        CompletedQuests = new Dictionary<uint, CompletedQuest>();
         _removed = new List<uint>();
     }
 
@@ -493,7 +493,7 @@ public class CharacterQuests
                 while (reader.Read())
                 {
                     var quest = new CompletedQuest();
-                    quest.Id = reader.GetUInt16("id");
+                    quest.Id = reader.GetUInt32("id");
                     quest.Body = new BitArray((byte[])reader.GetValue("data"));
                     CompletedQuests.Add(quest.Id, quest);
                 }
